Fixed user locks and semaphores
Also implemented a new way to allocate data for use by a process. Now,
each process has it's own "heap" (this is NOT the same as the heap in
userspace!!!) on the kernel that it can allocate data on. This special
"heap" is simply a linked list of kernel-space pages which only keep
track of the total free memory size in the page and the pointer to the
bottom of that free memory. When allocating on this "heap", the kernel
searches through all the pages for a page with enough free space. If
such a page is found, the page's base pointer and size are changed
accordingly and the pointer to the new allocated space is returned. If
no page is found, a new one is allocated with palloc_get_page(0) and
initilized accordingly. This system only frees the allocated data when
the process exits, so it is only suitable for allocations which are only supposed to be freed when the process exits (like user locks and user semaphores).